refactor : main, splash 에 해당하는 feature 패키지로 변경
