+++ Taehyung/AndroidStudy/features/Delivery/src/main/java/com/terry/delivery/features/main/DeliveryMainActivity.kt
@@ -1,4 +1,4 @@
-package com.terry.delivery.ui
+package com.terry.delivery.features.main
 
 import android.os.Bundle
 import com.terry.delivery.base.BaseActivity
