+++ Taehyung/AndroidStudy/features/Delivery/src/main/res/layout/activity_delivery_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.DeliveryMainActivity">
+    tools:context=".features.main.DeliveryMainActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_view_delivery_main"
