+++ Taehyung/AndroidStudy/features/Delivery/src/main/res/navigation/nav_main.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/splashFragment"
-        android:name="com.terry.delivery.ui.SplashFragment"
+        android:name="com.terry.delivery.features.splash.SplashFragment"
         android:label="SplashFragment"
         tools:layout="@layout/fragment_splash">
         <action
@@ -21,7 +21,7 @@
 
     <fragment
         android:id="@+id/mainFragment"
-        android:name="com.terry.delivery.ui.MainFragment"
+        android:name="com.terry.delivery.features.main.MainFragment"
         android:label="MainFragment"
         tools:layout="@layout/fragment_main">
 
@@ -63,19 +63,19 @@
 
     <fragment
         android:id="@+id/favoriteFragment"
-        android:name="com.terry.delivery.ui.FavoriteFragment"
+        android:name="com.terry.delivery.features.main.FavoriteFragment"
         android:label="FavoriteFragment"
         tools:layout="@layout/fragment_favorite" />
 
     <fragment
         android:id="@+id/orderListFragment"
-        android:name="com.terry.delivery.ui.OrderListFragment"
+        android:name="com.terry.delivery.features.main.OrderListFragment"
         android:label="OrderListFragment"
         tools:layout="@layout/fragment_order_list" />
 
     <fragment
         android:id="@+id/myPageFragment"
-        android:name="com.terry.delivery.ui.MyPageFragment"
+        android:name="com.terry.delivery.features.main.MyPageFragment"
         android:label="@string/my_page"
         tools:layout="@layout/fragment_my_page">
